KS-1-4 изменил тип переменной, чтобы в большом числе не съедало цифры.

+++ src/main/kotlin/lesson1_task4.kt
@@ -3,9 +3,9 @@ package org.example
 fun main() {
     var length: Long = 40868600000
     var age: Byte = 27
-    var partOfDay: Double = 0.075
+    var partOfDay: Float = 0.075f
     var seconds: Short = 6480
-    var partOfYear: Float = 2.0547945205479453E-4f
+    var partOfYear: Double = 2.0547945205479453E-4
     var apogee: Int = 327000
     println("Расстояние: $length")
     println("Возраст: $age")
